I think PCA almost works, again
