+++ covariance/CMakeLists.txt
@@ -1,5 +1,6 @@
 set(HEADERS
     ThrowParms.h
+    CovarianceUtils.h
     covarianceBase.h
     covarianceOsc.h
     covarianceXsec.h
@@ -7,6 +8,7 @@ set(HEADERS
 
 add_library(Covariance SHARED
     ThrowParms.cpp
+    CovarianceUtils.cpp
     covarianceBase.cpp
     covarianceOsc.cpp
     covarianceXsec.cpp
